update: created controlller files and routes for transactions

+++ src/migrations/20250923134440-create-users.js
@@ -29,7 +29,7 @@ module.exports = {
         allowNull: false,
       },
       wallet_balance: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(15, 2), 
         allowNull: false,
         defaultValue: 0.00
       },
